// packages/mobile/src/types.ts
export interface colorProps {
  whiteYellow: string;
  P: string;
  dropDownText: string;
  dropDownBack: string;
  dropDownSelectBack: string;
  tabActive: string;
  tabInActive: string;
  certificateHeaderBack: string;
  certificateHeader: string;
  certiSubHeaderBack: string;
  certiSubHeaderTitle: string;
  assessPointBackground: string;
  assessPointText: string;
  assessmentTitle: string;
  CircularProgressInactive: string;
  cardBackground2: string;
  cardBackground: string;
  bottomBorder: string;
  headerBackground: string;
  MasterBackground: string;
  headertext: string;
  cardBorder: string;
  titleOrange: string;
  background: string;
  lightBlue3: string;
  black: string;
  black2: string;
  white: string;
  primary: string;
  secondary: string;
  error: string;
  success: string;
  warning: string;
  Blue_2: string;
  Blue_Theme: string;
  Grey_3: string;
  Grey_1: string;
  Grey_2: string;
  Grey_4: string;
  Card_Gradian: string;
  ORANGE: string;
  YELLOW: string;
  HOVER_ORANGE: string;
  Card_2: string;
  Card_3: string;
  Card_4: string;
  Notification_red: string;
  Light_blue: string;
  Blue_Gradient: string[];
  green: string;
  Feed_Gradient: string[];
  cyanBlue: string;
  purple_light: string;
  cream: string;
  tealGreen: string;
  brightYellow: string;
  CorrectGreen: string;
  accountlinkBack: string;
  BorderColor: string;
  dividerColor: string;
  ProfileColor: string;
  ResuleColor: string;
  achivementBackground: string;
  hilight: string
}
export interface themeProps {
  colors: colorProps;
  dark: boolean;
}